error: unify two classes
